fix: rename testScreenForNavigation and rename the branch
