+++ app/src/main/java/com/github/onlynotesswent/MainActivity.kt
@@ -19,8 +19,8 @@ import com.github.onlynotesswent.resources.C
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.github.onlynotesswent.ui.navigation.Route
 import com.github.onlynotesswent.ui.navigation.Screen
-import com.github.onlynotesswent.ui.testScreenForNavigation.AuthenticationScreen
-import com.github.onlynotesswent.ui.testScreenForNavigation.OverviewScreen
+import com.github.onlynotesswent.ui.screen.AuthenticationScreen
+import com.github.onlynotesswent.ui.screen.OverviewScreen
 import com.github.onlynotesswent.ui.theme.SampleAppTheme
 
 class MainActivity : ComponentActivity() {
